fix: circuit tests and tsconfigs (#621)
* fix: circuit tests and tsconfigs

* fix-ci: downgrade typescript version

* fix: yarn.lock

* fix: linting in common
