+++ circuits/package.json
@@ -19,7 +19,7 @@
     "lint": "prettier --check .",
     "nice": "prettier --write .",
     "test": "yarn test-base 'tests/**/*.test.ts' --exit",
-    "test-base": "NODE_OPTIONS='--loader ts-node/esm' yarn ts-mocha --max-old-space-size=8192",
+    "test-base": "yarn ts-mocha -n import=tsx --max-old-space-size=8192 --paths -p tsconfig.spec.json",
     "test-custom-hasher": "yarn test-base 'tests/other_circuits/custom_hasher.test.ts' --exit",
     "test-disclose": "yarn test-base 'tests/disclose/vc_and_disclose.test.ts' --exit",
     "test-disclose-id": "yarn test-base 'tests/disclose/vc_and_disclose_id.test.ts' --exit",
@@ -80,6 +80,7 @@
     "mocha": "^10.3.0",
     "prettier": "^3.3.3",
     "ts-mocha": "^10.0.0",
-    "ts-node": "^10.9.2"
+    "tsconfig-paths": "^4.2.0",
+    "tsx": "^4.20.3"
   }
 }
