+++ circuits/tests/register_id/register_id.test.ts
@@ -12,8 +12,11 @@ import { poseidon6 } from 'poseidon-lite';
 import { hashAlgosTypes, ID_CARD_ATTESTATION_ID } from '@selfxyz/common/constants/constants';
 import { parseCertificateSimple } from '@selfxyz/common/utils/certificate_parsing/parseCertificateSimple';
 import serialized_dsc_tree from '../../../common/pubkeys/serialized_dsc_tree.json' with { type: 'json' };
-import { genMockIdDoc } from '@selfxyz/common/utils/passports/genMockIdDoc';
+import { genMockIdDocAndInitDataParsing } from '@selfxyz/common/utils/passports/genMockIdDoc';
+import { fileURLToPath } from 'url';
 dotenv.config();
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const testSuite = process.env.FULL_TEST_SUITE === 'true' ? fullSigAlgs : sigAlgs;
 
@@ -33,7 +36,7 @@ testSuite.forEach(
       this.timeout(0);
       let circuit: any;
 
-      const passportData = genMockIdDoc({
+      const passportData = genMockIdDocAndInitDataParsing({
         idType: 'mock_id_card',
         dgHashAlgo: dgHashAlgo as hashAlgosTypes,
         eContentHashAlgo: eContentHashAlgo as hashAlgosTypes,
@@ -56,9 +59,9 @@ testSuite.forEach(
           ),
           {
             include: [
-              'node_modules',
-              './node_modules/@zk-kit/binary-merkle-root.circom/src',
-              './node_modules/circomlib/circuits',
+              '../node_modules',
+              '../node_modules/@zk-kit/binary-merkle-root.circom/src',
+              '../node_modules/circomlib/circuits',
             ],
           }
         );
