+++ circuits/tests/utils/rsapss.test.ts
@@ -7,6 +7,9 @@ import {
 } from './generateMockInputsRsaPss.js';
 import { expect } from 'chai';
 import { fullAlgorithms, sigAlgs, AdditionalCases } from './testcase/rsapss.js';
+import { fileURLToPath } from 'url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 describe('VerifyRsapss Circuit Test', function () {
   this.timeout(0);
@@ -27,7 +30,7 @@ describe('VerifyRsapss Circuit Test', function () {
             `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
           ),
           {
-            include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+            include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
           }
         );
 
@@ -65,7 +68,7 @@ describe('VerifyRsapss Circuit Test', function () {
           `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
         ),
         {
-          include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+          include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
         }
       );
 
@@ -98,7 +101,7 @@ describe('VerifyRsapss Circuit Test', function () {
           `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
         ),
         {
-          include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+          include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
         }
       );
 
@@ -130,7 +133,7 @@ describe('VerifyRsapss Circuit Test', function () {
           `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
         ),
         {
-          include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+          include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
         }
       );
 
@@ -160,7 +163,7 @@ describe('VerifyRsapss Circuit Test', function () {
           `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
         ),
         {
-          include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+          include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
         }
       );
 
@@ -188,7 +191,7 @@ describe('VerifyRsapss Circuit Test', function () {
           `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
         ),
         {
-          include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+          include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
         }
       );
 
@@ -219,7 +222,7 @@ describe('VerifyRsapss Circuit Test', function () {
           `../../circuits/tests/utils/rsapss/test_${algorithm.algo}_${algorithm.saltLength}.circom`
         ),
         {
-          include: ['node_modules', './node_modules/@zk-kit/binary-merkle-root.circom/src'],
+          include: ['../node_modules', '../node_modules/@zk-kit/binary-merkle-root.circom/src'],
         }
       );
